+++ configs/_base_/models/tpn_tsm_r50.py
@@ -33,4 +33,4 @@ model = dict(
         init_std=0.01),
     # model training and testing settings
     train_cfg=None,
-    test_cfg=dict(average_clips=None))
+    test_cfg=dict(average_clips='prob', fcn_test=True))