+++ mmaction/models/recognizers/recognizer2d.py
@@ -1,3 +1,6 @@
+import torch
+from torch import nn
+
 from ..registry import RECOGNIZERS
 from .base import BaseRecognizer
 
@@ -15,6 +18,14 @@ class Recognizer2D(BaseRecognizer):
         losses = dict()
 
         x = self.extract_feat(imgs)
+
+        if self.backbone_from == 'torchvision':
+            if len(x.shape) == 4 and (x.shape[2] > 1 or x.shape[3] > 1):
+                # apply adaptive avg pooling
+                x = nn.AdaptiveAvgPool2d(1)(x)
+            x = x.reshape((x.shape[0], -1))
+            x = x.reshape(x.shape + (1, 1))
+
         if hasattr(self, 'neck'):
             x = [
                 each.reshape((-1, num_segs) +
@@ -37,11 +48,18 @@ class Recognizer2D(BaseRecognizer):
         """Defines the computation performed at every call when evaluation,
         testing and gradcam."""
         batches = imgs.shape[0]
-
         imgs = imgs.reshape((-1, ) + imgs.shape[2:])
         num_segs = imgs.shape[0] // batches
 
         x = self.extract_feat(imgs)
+
+        if self.backbone_from == 'torchvision':
+            if len(x.shape) == 4 and (x.shape[2] > 1 or x.shape[3] > 1):
+                # apply adaptive avg pooling
+                x = nn.AdaptiveAvgPool2d(1)(x)
+            x = x.reshape((x.shape[0], -1))
+            x = x.reshape(x.shape + (1, 1))
+
         if hasattr(self, 'neck'):
             x = [
                 each.reshape((-1, num_segs) +
@@ -68,9 +86,49 @@ class Recognizer2D(BaseRecognizer):
 
         return cls_score
 
+    def _do_fcn_test(self, imgs):
+        # [N, num_crops * num_segs, C, H, W] ->
+        # [N * num_crops * num_segs, C, H, W]
+        batches = imgs.shape[0]
+        imgs = imgs.reshape((-1, ) + imgs.shape[2:])
+        num_segs = self.test_cfg.get('num_segs', self.backbone.num_segments)
+
+        if self.test_cfg.get('flip', False):
+            imgs = torch.flip(imgs, [-1])
+        x = self.extract_feat(imgs)
+
+        if hasattr(self, 'neck'):
+            x = [
+                each.reshape((-1, num_segs) +
+                             each.shape[1:]).transpose(1, 2).contiguous()
+                for each in x
+            ]
+            x, _ = self.neck(x)
+        else:
+            x = x.reshape((-1, num_segs) +
+                          x.shape[1:]).transpose(1, 2).contiguous()
+
+        # When using `TSNHead` or `TPNHead`, shape is [batch_size, num_classes]
+        # When using `TSMHead`, shape is [batch_size * num_crops, num_classes]
+        # `num_crops` is calculated by:
+        #   1) `twice_sample` in `SampleFrames`
+        #   2) `num_sample_positions` in `DenseSampleFrames`
+        #   3) `ThreeCrop/TenCrop/MultiGroupCrop` in `test_pipeline`
+        #   4) `num_clips` in `SampleFrames` or its subclass if `clip_len != 1`
+        cls_score = self.cls_head(x, fcn_test=True)
+
+        assert cls_score.size()[0] % batches == 0
+        # calculate num_crops automatically
+        cls_score = self.average_clip(cls_score,
+                                      cls_score.size()[0] // batches)
+        return cls_score
+
     def forward_test(self, imgs):
         """Defines the computation performed at every call when evaluation and
         testing."""
+        if self.test_cfg.get('fcn_test', False):
+            # If specified, spatially fully-convolutional testing is performed
+            return self._do_fcn_test(imgs).cpu().numpy()
         return self._do_test(imgs).cpu().numpy()
 
     def forward_dummy(self, imgs):